fix : changes during audit use float 64 as return values in average and variance for acuracy during calculation of variance and standard deviation

+++ statcalc/deviation.go
@@ -2,6 +2,6 @@ package statcalc
 
 import "math"
 
-func CalculateStandardDeviation(variance int) int {
+func CalculateStandardDeviation(variance float64) int {
 	return int(math.Round(math.Sqrt(float64(variance))))
 }
